+++ zhmccli/_cmd_capacitygroup.py
@@ -89,6 +89,13 @@ def capacitygroup_show(cmd_ctx, cpc, capacitygroup):
     """
     Show the details of a capacity group.
 
+    The following properties are shown in addition to those returned by the HMC:
+
+    \b
+      - 'parent-name' - Name of the parent CPC.
+      - 'partition-names' - Names of the partitions referenced by
+        'partition-uris' (index-correlated).
+
     In addition to the command-specific options shown in this help text, the
     general options (see 'zhmc --help') can also be specified right after the
     'zhmc' command name.
@@ -299,7 +306,19 @@ def cmd_capacitygroup_show(cmd_ctx, cpc_name, capacitygroup_name):
     except zhmcclient.Error as exc:
         raise click_exception(exc, cmd_ctx.error_format)
 
-    print_properties(cmd_ctx, capacitygroup.properties, cmd_ctx.output_format)
+    properties = dict(capacitygroup.properties)
+
+    # Add artificial property 'parent-name'
+    properties['parent-name'] = cpc_name
+
+    # Add artificial property 'partition-names'
+    part_names = []
+    for part_uri in capacitygroup.properties['partition-uris']:
+        part_props = client.session.get(part_uri)
+        part_names.append(part_props['name'])
+    properties['partition-names'] = part_names
+
+    print_properties(cmd_ctx, properties, cmd_ctx.output_format)
 
 
 def cmd_capacitygroup_update(cmd_ctx, cpc_name, capacitygroup_name, options):