+++ zhmccli/_cmd_storagevolume.py
@@ -102,6 +102,13 @@ def storagevolume_show(cmd_ctx, storagegroup, storagevolume):
     """
     Show the details of a storage volume.
 
+    The following properties are shown in addition to those returned by the HMC:
+
+    \b
+      - 'parent-name' - Name of the parent Storage Group.
+      - 'adapter-name' - Name of Adapter referenced by 'adapter-uri',
+        if present (for NVME-based volumes).
+
     In addition to the command-specific options shown in this help text, the
     general options (see 'zhmc --help') can also be specified right after the
     'zhmc' command name.
@@ -318,7 +325,22 @@ def cmd_storagevolume_show(cmd_ctx, stogrp_name, stovol_name):
     except zhmcclient.Error as exc:
         raise click_exception(exc, cmd_ctx.error_format)
 
-    print_properties(cmd_ctx, stovol.properties, cmd_ctx.output_format)
+    properties = dict(stovol.properties)
+
+    # Add artificial property 'parent-name'
+    properties['parent-name'] = stogrp_name
+
+    # Add artificial property 'adapter-name'
+    try:
+        adapter_uri = stovol.get_property('adapter-uri')
+    except KeyError:
+        pass
+    else:
+        # The storage volume is NVME type (so it has an adapter)
+        adapter_props = client.session.get(adapter_uri)
+        properties['adapter-name'] = adapter_props['name']
+
+    print_properties(cmd_ctx, properties, cmd_ctx.output_format)
 
 
 def cmd_storagevolume_create(cmd_ctx, stogrp_name, options):